Move ballerina code generating logic to file builder
