+++ stdlib/grpc/src/main/java/org/ballerinalang/net/grpc/builder/utils/BalGenConstants.java
@@ -55,9 +55,9 @@ public class BalGenConstants {
     
     public static final String PACKAGE_SEPARATOR_REGEX = "\\.";
     
-    static final String TEMPLATES_SUFFIX = ".mustache";
+    public static final String TEMPLATES_SUFFIX = ".mustache";
     
-    static final String TEMPLATES_DIR_PATH_KEY = "templates.dir.path";
+    public static final String TEMPLATES_DIR_PATH_KEY = "templates.dir.path";
     
     private static final String DEFAULT_TEMPLATE_DIR = RESOURCE_SEPARATOR + "templates";
